updated brewfile and makesymlinks file

+++ makesymlinks.sh
@@ -1,31 +1,34 @@
 #!/bin/bash 
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old             # old dotfiles backup directory
-files="vimrc vim zshrc tmux.conf zimrc yabairc skhdrc"    # list of files/folders to symlink in homedir
+files="vimrc vim zshrc tmux.conf zimrc yabairc skhdrc alacritty.yml z zim zlogin zshenv"    # list of files/folders to symlink in homedir
 
 ##########
 
-# create dotfiles_old in homedir
-echo -n "Creating $olddir for backup of any existing dotfiles in ~ ..."
-mkdir -p $olddir
-echo "done"
+echo "Make sure you have run your brewfile first"
+read -p "Do you want to continue" -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+        # do dangerous stuff
+    # create dotfiles_old in homedir
+    echo -n "Creating $olddir for backup of any existing dotfiles in ~ ..."
+    mkdir -p $olddir
+    echo "done"
 
-# change to the dotfiles directory
-echo -n "Changing to the $dir directory ..."
-cd $dir
-echo "done"
+    # change to the dotfiles directory
+    echo -n "Changing to the $dir directory ..."
+    cd $dir
+    echo "done"
 
-# move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
-for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
-done
+    # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
+    for file in $files; do
+        echo "Moving any existing dotfiles from ~ to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+        echo "Creating symlink to $file in home directory."
+        ln -s $dir/$file ~/.$file
+    done
+fi
 
 
-# Install Vim via brew + vundle
-# install zim 
-# install Z
-# install nvm 
-# install go
+
